makefile && FLags updated

+++ source/libft/check/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: ebennace <ebennace@student.42lausanne.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/13 18:07:53 by ebennace          #+#    #+#             */
-/*   Updated: 2022/08/10 10:54:07 by ebennace         ###   ########.fr       */
+/*   Updated: 2022/10/02 16:24:12 by ebennace         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../libft.h"
 
-size_t	ft_strlen(const char *str)
+int	ft_strlen(const char *str)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (str[i] != '\0')
